+++ src/screens/blocks/components/list/types.ts
@@ -1,12 +1,3 @@
-export type BlockType = {
-  block: number; // round
-  timestamp: number;
-  txs: number;
-  shard: number;
-  size: number;
-  hash: string;
-}
-
 export type BlockState = {
   page: number;
   loading: boolean;