+++ src/screens/validator_details/components/index.ts
@@ -1,0 +1,13 @@
+import Profile from './profile';
+import Stake from './stake';
+import Overview from './overview';
+import ContractDetails from './contract_details';
+import Nodes from './nodes';
+
+export {
+  Profile,
+  Stake,
+  Overview,
+  ContractDetails,
+  Nodes,
+};